+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-02-04
+### Added
+- user friendly error string on cpc_init
+- command to get bootloader version running on RCP target
+- command to get the app version from the Application_Properties_t struct of the RCP application
+
 ## [0.1.1] - 2024-01-16
 ### Changed 
 - Adds detail to the documentation for components required to enable SWO debug